auth_api/python/requirements.txt
Flask
pyjwt
SQLAlchemy
pymysql
